+++ sdk/azcore/runtime/policy_bearer_token_test.go
@@ -246,3 +246,15 @@ func TestCheckHTTPSForAuth(t *testing.T) {
 	require.NoError(t, err)
 	require.NoError(t, checkHTTPSForAuth(req))
 }
+
+func TestBearerTokenPolicy_NilCredential(t *testing.T) {
+	policy := NewBearerTokenPolicy(nil, nil, nil)
+	pl := exported.NewPipeline(shared.TransportFunc(func(req *http.Request) (*http.Response, error) {
+		require.Zero(t, req.Header.Get(shared.HeaderAuthorization))
+		return &http.Response{}, nil
+	}), policy)
+	req, err := NewRequest(context.Background(), "GET", "http://contoso.com")
+	require.NoError(t, err)
+	_, err = pl.Do(req)
+	require.NoError(t, err)
+}